pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orglens"
version = "0.1.0"
description = "Organizational lens for AI agents — topology, grammar, state"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
orglens = "orglens.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
